compiler: move Config struct to compileopts
